Add QgsDataItem::Fast flag for QgsPGLayerItem
This prevents creating Futures for building GUI elements
for each layer attribute in parallel.

If QGis is quit with generic Geometry layers expanded in the PostGIS
browser, then restarting QGis and reopening these schemas in the browser
results often in never ending hourglasses shown for the Geometry layers.

The root cause is that some Futures are never completed, probably
due to a race condition. Adding the Fast flag prevents creating these
Futures and does not result in a measurable slowdown for the GUI.